refactor: firebase core를 firestore로 분리
